flask/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
PyJWT==2.8.0
PyMySQL==1.1.0